+++ domain/src/main/kotlin/gr/eduinvoice/domain/lesson/UpdateLessonInvoicedStatus.kt
@@ -6,6 +6,6 @@ import javax.inject.Inject
 class UpdateLessonInvoicedStatus @Inject constructor(
     private val dao: LessonDao
 ) {
-    suspend operator fun invoke(ids: List<Long>, invoiced: Boolean) =
-        dao.updateInvoicedStatus(ids, invoiced)
+    suspend operator fun invoke(ids: List<Long>, invoiced: Boolean, userId: Long) =
+        dao.updateInvoicedStatus(ids, invoiced, userId)
 }