+++ domain/src/main/kotlin/gr/eduinvoice/domain/lesson/UpdateLessonPaidStatus.kt
@@ -6,5 +6,6 @@ import javax.inject.Inject
 class UpdateLessonPaidStatus @Inject constructor(
     private val dao: LessonDao
 ) {
-    suspend operator fun invoke(ids: List<Long>, paid: Boolean) = dao.updatePaidStatus(ids, paid)
+    suspend operator fun invoke(ids: List<Long>, paid: Boolean, userId: Long) =
+        dao.updatePaidStatus(ids, paid, userId)
 }